pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datagpu"
version = "0.1.0"
description = "Open-source data compiler for AI training datasets"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "DataGPU Contributors"}
]
keywords = ["machine-learning", "data-processing", "dataset", "compiler", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "polars>=0.19.0",
    "pyarrow>=14.0.0",
    "xxhash>=3.4.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
embeddings = [
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
]
distributed = [
    "ray>=2.7.0",
]

[project.scripts]
datagpu = "datagpu.cli:app"

[project.urls]
Homepage = "https://github.com/datagpu/datagpu"
Repository = "https://github.com/datagpu/datagpu"
Documentation = "https://github.com/datagpu/datagpu#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["datagpu*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=datagpu --cov-report=term-missing"
